Refactored all of main.js code
