--- unpacked-app/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "alloy-demo-app",
-  "version": "1.0.1",
-  "description": "Alloy Journey Application Demo App",
-  "main": "main.js",
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@alloyidentity/web-sdk": "^2.3.1",
-    "axios": "^1.6.2",
-    "body-parser": "^1.20.2",
-    "cors": "^2.8.5",
-    "dotenv": "^16.3.1",
-    "electron-is-dev": "^2.0.0",
-    "electron-store": "^8.1.0",
-    "express": "^4.18.2",
-    "express-rate-limit": "^7.1.5",
-    "helmet": "^7.1.0",
-    "node-fetch": "^3.3.2",
-    "shepherd.js": "^14.5.0",
-    "winston": "^3.10.0",
-    "winston-daily-rotate-file": "^4.7.1"
-  }
-}
