--- unpacked-app/public/sample-files/sample-config.json
@@ -1,7 +1,0 @@
-{
-  "ALLOY_TOKEN": "your-alloy-token-here",
-  "ALLOY_SECRET": "your-alloy-secret-here",
-  "ALLOY_SDK_KEY": "your-sdk-key-here",
-  "ALLOY_JOURNEY_TOKEN": "your-journey-token-here",
-  "ALLOY_BASE_URL": "https://sandbox.alloy.co"
-}
